Move function to utils, some changes for LLama3
