+++ utils/text.py
@@ -1,3 +1,4 @@
+import datetime
 import string
 from rouge_score import rouge_scorer
 
@@ -16,3 +17,21 @@ def truncate(s: str, length: int = 40) -> str:
     if len(s) <= length:
         return s
     return s[:length - 3].rstrip(string.punctuation) + "..."
+
+
+def td_format(td: datetime.timedelta) -> str:
+    seconds = int(td.total_seconds())
+    periods = [
+        ('year', 3600*24*365), ('month', 3600*24*30), ('day', 3600*24), ('hour', 3600), ('minute', 60), ('second', 1)
+    ]
+    parts = list()
+    for period_name, period_seconds in periods:
+        if seconds > period_seconds:
+            period_value, seconds = divmod(seconds, period_seconds)
+            has_s = 's' if period_value > 1 else ''
+            parts.append("%s %s%s" % (period_value, period_name, has_s))
+    if len(parts) == 0:
+        return "just now"
+    if len(parts) == 1:
+        return f"{parts[0]} ago"
+    return " and ".join([", ".join(parts[:-1])] + parts[-1:]) + " ago"
